nbaNews/requirements.txt
beautifulsoup4==4.7.1
certifi==2018.11.29
chardet==3.0.4
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.1.7
django-rest-framework==0.1.0
djangorestframework==3.9.2
gunicorn==19.9.0
idna==2.8
psycopg2==2.7.7
pytz==2018.9
requests==2.21.0
soupsieve==1.8
static3==0.7.0
urllib3==1.24.1
